ms code

+++ module2-boot/ractive_boot/orderapp2/src/main/java/com/orderapp/orderapp/repository/OrderRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.r2dbc.repository.R2dbcRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OrderRepo extends R2dbcRepository<Order,Long> {
+public interface OrderRepo extends R2dbcRepository<Order,Integer> {
 }
